Fix DSV parse bug where last field wasn't recorded
The internal validation function elz_hlp_is_valid_char invalidates a
NULL byte, which the DSV parser uses as a delimiter trigger to add the
last field to the set. With this in mind, the check is now performed
in the nonescape logic rather than the main parser, the idea being that
all other characters (delimiter and escape characters and codes) will
have all been validated by the configuration functions before parsing
begins.